% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare}
\alias{prepare}
\title{prepare Function}
\usage{
prepare(explicit, iat, sessions, tasks, demo, type)
}
\arguments{
\item{explicit}{A string for the path of explicit.txt file or the equivalent.}

\item{iat}{A string for the path of iat.txt file or the equivalent.}

\item{sessions}{A string for the path of sessions.txt file or the equivalent.}

\item{tasks}{A string for the path of sessionTasks.txt file or the equivalent.}

\item{demo}{A string for the path of demographics.txt file or the equivalent.}

\item{type}{A variable indicating data collection procedure. Possible value: mTurk, PI, PIMH}
}
\value{
To Be Specified.
}
\description{
\code{sum}
}
\details{
This is the main function of GMDPP package. It will take all the standard files you will get from PI or MTurk, transform them into formats that required by PI reseseach team(Insert link here). I t will also return necessary variables that include information that you will find handy in later data analysis.
}
\examples{
 \dontrun{
     prepare('~/explicit.txt','~/iat.txt','~/sessions.txt','~/sessionTasks.txt','~/demographics.txt', 'mTurk')
   }
}
\seealso{
\code{\link{~/Requirements}} For files naming consistency and dataset cleaning requirements.
}
\keyword{Data}
\keyword{pre-analysis}

