% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.PI}
\alias{print.PI}
\title{print.PI Function}
\usage{
\method{print}{PI}(dataPack)
}
\arguments{
\item{rawPack}{A PI object, included subfields like Explicit, Sessions, Tasks, IAT and Demo which representing the original five datasets(required to process) and further addon variables(optional).}
}
\value{
To Be Specified.
}
\description{
Print method for PI objects.
}
\details{
This is a small function to help print the PI object more effectively.
}
\examples{
 \dontrun{
		rawData <- takeFive('~/explicit.txt','~/iat.txt','~/sessions.txt','~/sessionTasks.txt','~/demographics.txt','mTurk')
     prepare(rawData)
   }
}
\seealso{
\code{\link{~/Requirements}} For files naming consistency and dataset cleaning requirements.
}
\keyword{datasets}
\keyword{print}

