% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/say_hello.R
\name{say_hello}
\alias{say_hello}
\title{Say Hello Function}
\usage{
say_hello(say = TRUE)
}
\arguments{
\item{say}{a dummy variable, has no influence on the program}
}
\description{
Relax, this function just want to sing a song for you. If you happen to know what song it is singing, shoot Diheng an email! He will take you out for a coffee.
}
\examples{
say_hello()
}
\keyword{hello}

