% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.PI}
\alias{summary.PI}
\title{Summary method Function}
\usage{
\method{summary}{PI}(dataPack)
}
\arguments{
\item{rawPack}{A PI object, included subfields like Explicit, Sessions, Tasks, IAT and Demo which representing the original five datasets(required to process) and further addon variables(optional).}
}
\value{
To Be Specified.
}
\description{
\code{sum}
}
\details{
This is the main function of GMDPP package. It will take the raw data object
you get from the takeFive, transform them into formats that required
by PI reseseach team(Insert link here). It will also return necessary variables
that include information that you will find handy in later data cleaning.
}
\examples{
 \dontrun{
		rawData <- takeFive('~/explicit.txt','~/iat.txt','~/sessions.txt','~/sessionTasks.txt','~/demographics.txt','mTurk')
     prepare(rawData)
   }
}
\seealso{
\code{\link{~/Requirements}} For files naming consistency and dataset cleaning requirements.
}
\keyword{Data}
\keyword{pre-analysis}

