% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare}
\alias{prepare}
\title{prepare Function}
\usage{
prepare(rawPack)
}
\arguments{
\item{rawPack}{A PI object, included elements like Explicit, Sessions, Tasks, IAT and Demo which representing the original five datasets(required to process) and further addon variables(optional).}
}
\value{
If it works correctly, there will be an objects return as original datasets and extra elements for reference. They are: $Explicit, $IAT, $Sessions, $Tasks and $Demo for actual datasets, and $participant_id $participatedNum $completedID $completedNum etc.
}
\description{
\code{Basic handling for PI raw datasets.}
}
\details{
This is one of the main function of GMDPP package. It will take the raw data object
you get from the takeFive, transform them into formats that required
by PI reseseach team(Insert link here). It will also return necessary variables
that include information that you will find handy in later data cleaning.
}
\examples{
 \dontrun{
		rawData <- takeFive('~/explicit.txt','~/iat.txt','~/sessions.txt','~/sessionTasks.txt','~/demographics.txt','mTurk')
     prepare(rawData)
   }
}
\seealso{
\code{\link{~/Requirements}} For files naming consistency and dataset cleaning requirements.
}
\keyword{Data}
\keyword{pre-analysis}

