% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.PI}
\alias{summary.PI}
\title{summary Function}
\usage{
\method{summary}{PI}(dataPack)
}
\arguments{
\item{rawPack}{A PI object, included subfields like Explicit, Sessions, Tasks, IAT and Demo which representing the original five datasets(required to process) and further addon variables(optional).}
}
\value{
To Be Specified.
}
\description{
Summary method for PI objects.
}
\details{
This is a small function to summary main information out of a PI object.
}
\examples{
 \dontrun{
		rawData <- takeFive('~/explicit.txt','~/iat.txt','~/sessions.txt','~/sessionTasks.txt','~/demographics.txt','mTurk')
     prepare(rawData)
   }
}
\seealso{
\code{\link{~/Requirements}} For files naming consistency and dataset cleaning requirements.
}
\keyword{datasets}
\keyword{of}
\keyword{summary}

